+++ src/components/WorkflowObserver/WorkflowObserver.js
@@ -7,14 +7,10 @@ import { env } from 'config'
 import Pusher from 'pusher-js'
 
 import { mergedTheme } from 'theme'
+import LargeMessageBox from 'components/LargeMessageBox'
 
 const StyledAnchor = styled(Anchor)`
   text-decoration: none;
-`
-
-const LargeBox = styled(Box)`
-  min-width: 50%;
-  min-height: 10em;
 `
 
 let pusher = null
@@ -85,16 +81,9 @@ const WorkflowObserver = function ({
   return (
     <Box wrap={true} direction="row">
       {(recentSubjects.length === 0) &&
-        <LargeBox
-          background='#e5e5e5'
-          align='center'
-          justify='center'
-          margin={{ vertical: 'small', horizontal: 'auto', }}
-          round='xsmall'
-          size='medium'
-        >
+        <LargeMessageBox>
           <Text>No Classifications yet.</Text>
-        </LargeBox>
+        </LargeMessageBox>
       }
       {recentSubjects.map((subject, index) => (
         <StyledAnchor
